requirements/local.txt
-r base.txt
django-debug-toolbar==6.1.0
django-extensions==3.2.3
